legacy_python_agent/pyproject.toml
[project]
name = "sentinel-ai"
version = "0.1.0"
description = "AI-Powered Server & Website Security Platform"
authors = [
    {name = "SentinelAI Team", email = "team@sentinelai.local"},
]
dependencies = [
    "click>=8.1.7",
    "pydantic>=2.6.1",
    "pyyaml>=6.0.1",
    "aiosqlite>=0.19.0",
    "watchdog>=4.0.0",
    "apscheduler>=3.10.4",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "scikit-learn>=1.4.0",  # For anomaly detection
    "numpy>=1.26.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "AGPL-3.0"}

[project.scripts]
sentinelctl = "cli.sentinel_cli.main:cli"
sentinel-agent = "agent.sentinel.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent/sentinel", "cli/sentinel_cli"]
